feat: add hierarchical application config (#10)
Makes use of [`config-rs`](https://github.com/mehcode/config-rs) to
implement hierarchical configuration from files and environment
variables. This makes `clap` obsolete.

In the future we'll have more configuration options, e.g. when we add
calendar support for events. Also we need to be able to configure things
canonical URLs (for SEO) which are only relevant in production.