feat(calendar): add `EventSource` abstraction (#12)
Adds an `EventSource` trait which allows the implementation of multiple
sources for calendar events.

Right now only `StaticEventSource` was implemented to use a static list
of events from the config. This is mostly useful during testing and
development.

Later, we want to also add an `EventSource` for the Google Calendar
which should be used in production.

Please note that the API for the `EventSource` trait is just an initial
draft. We can obviously change it once we see what's needed by Google
Calendar. I already designed it to be `async` though, because we'll need
that later.