chore: add support for canonical `link` tag
If a `canonical_url` is configured in the `site` config, the `link` tag
with the canonical URL will be added to the HTML head.

This is only relevant for SEO, here's what Google has to say about this:
https://developers.google.com/search/docs/crawling-indexing/consolidate-duplicate-urls

I don't like the way we pass down the `request_path` into the template.
In the best case I'd like to have this done implicitly somehow.
Otherwise we will need to do this for every endpoint and it's easy to
forget.