feat(events): fetch Google Calendar events in the background
Fetching events via Google Calendar API is slow and takes ~3 seconds.
This negatively impacts page load performance.

This moves the heavy event fetching into a background task which is
triggered when the homepage is visited, updating the event cache for the
next page load. This makes all page loads very fast, except for the
first one right after the application starts since it needs to populate
the event cache in the foreground.