feat(calendar): switch to periodic sync (#48)
We're periodically fetching all events from the event source instead of
smaller ranges. This makes quite a bit of code and the whole caching
layer obsolete.

It also avoids the first request to the homepage on a given day from
taking too long, because there's no such thing as cache misses anymore.