feat: add `events` pages and limit event display on homepage
The newly added `/events` page will display all upcoming events from the
next 12 months, while on the homepage we now only display events for the
next 3 months and add a link to the events page, so users can discover
the longer list if they need.

Also, we do not display events of the past month anymore because it was
useless and confusing.