feat: add support for event descriptions (#280)
* feat: add support for event descriptions

Really basic for now. Displays the description below the event date and
title, if available. The description is run through a markdown parser
and compiler to convert URLs in the description into clickable links.
Other shenanigans like bold text are also possible.

Another thing this change does is to update the overall font sizes. Big
screens now get bigger font and wider events table to increase
readability. We're getting old..

* fix: don't convert description

The description from the Google Calendar already is HTML, so we don't
need to process it any further.

* feat: properly handle a mix of markdown and HTML

With special handling for markdown's autolinks if a URL was already
wrapped in an `<a href="...">` element.

* fix: just remove empty anchors

Don't try to be smart.

* fix: improve description style